KA-137 - Move reducers operating items into separate folder
