+++ onboarding-task/src/reducers/ListItemReducers/itemIdsReducer.ts
@@ -3,9 +3,9 @@ import {
   ADD_ITEM, DELETE_ITEM,
   FETCH_ITEMS_SUCCESS,
   POST_ITEM_SUCCESS
-} from '../constants/actionTypes';
-import { IAction } from '../actionCreators/IAction';
-import { IItemData } from '../models/IItem';
+} from '../../constants/actionTypes';
+import { IAction } from '../../actionCreators/IAction';
+import { IItemData } from '../../models/IItem';
 
 const itemIdsReducer = (state: OrderedSet<string> = OrderedSet<string>(), action: IAction): OrderedSet<string> => {
   switch (action.type) {
