+++ onboarding-task/src/reducers/ListItemReducers/itemReducer.ts
@@ -1,13 +1,13 @@
-import { Item } from '../models/Item';
+import { Item } from '../../models/Item';
 import {
   ADD_ITEM,
   ENABLE_EDIT_ITEM,
   SAVE_CHANGES_TO_ITEM,
   CANCEL_CHANGES_TO_ITEM,
   POST_ITEM_SUCCESS,
-} from '../constants/actionTypes';
-import { IAction } from '../actionCreators/IAction';
-import { IItem, IItemData } from '../models/IItem';
+} from '../../constants/actionTypes';
+import { IAction } from '../../actionCreators/IAction';
+import { IItem, IItemData } from '../../models/IItem';
 
 const itemReducer = (state: IItem = new Item(), action: IAction): IItem => {
   switch (action.type) {
