+++ onboarding-task/src/reducers/rootReducer.ts
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux';
-import { itemsReducer } from './itemsReducer';
-import { itemIdsReducer } from './itemIdsReducer';
+import { itemsReducer } from './ListItemReducers/itemsReducer';
+import { itemIdsReducer } from './ListItemReducers/itemIdsReducer';
 import { fetchingFlagReducer } from './fetchingFlagReducer';
 import { errorReducer } from './errorReducer';
 import { postingFlagReducer } from './postingFlagReducer';
